requirements.txt
torch
transformers
openbmb
fastapi
uvicorn
pillow
peft
torchvision
python-multipart
runpod
